requirements.txt
annotated-types==0.7.0
anyio==4.12.1
azure-ai-agents==1.1.0
azure-ai-projects==2.0.1
azure-core==1.39.0
azure-identity==1.25.3
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.13.0
msal==1.35.1
msal-extensions==1.3.1
openai==2.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
